fix(music): accept string duration from flowmusic
flowmusic returns result.music[].duration as a quoted number (e.g. "181.7"), which broke unmarshaling into float64 and aborted polling. Add FlexFloat (accepts JSON number or numeric string) and use it for MusicTrack.Duration. Add a regression test.

+++ cmd/music_shared.go
@@ -104,8 +104,8 @@ func displayMusicResult(task *types.MusicTaskData) {
 			}
 			fmt.Printf("Track %d: %s\n", i+1, title)
 			switch {
-			case track.Duration > 0:
-				fmt.Printf("  Duration: %.1fs\n", track.Duration)
+			case float64(track.Duration) > 0:
+				fmt.Printf("  Duration: %.1fs\n", float64(track.Duration))
 			case track.DurationSeconds != "":
 				fmt.Printf("  Duration: %ss\n", track.DurationSeconds)
 			}
